requirements.txt
facebook-sdk==1.0.0
httplib2==0.9.2
oauthlib==1.1.2
python-dateutil==2.5.3
python-instagram==1.3.2
requests==2.10.0
requests-oauthlib==0.6.2
simplejson==3.8.2
six==1.10.0
tweepy==3.5.0
